+++ react/constants/Identifiers.ts
@@ -1,0 +1,6 @@
+export const enum IdentifierType {
+  PRODUCT_ID = 'ProductId',
+  PRODUCT_SKU_ITEM_ID = 'ProductSkuItemId',
+  PRODUCT_REFERENCE_ID = 'ProductReferenceId',
+  PRODUCT_SKU_REFERENCE_ID = 'ProductSkuReferenceId',
+}